+++ src/client/parse.rs
@@ -1305,4 +1305,12 @@ mod tests {
             _ => panic!(),
         };
     }
+
+    /// Tests parsing SDP from `the macro-video rtsp server`, with missing origin line.
+    #[test]
+    fn macrovideo() {
+        // DESCRIBE.
+        let p = parse_describe("rtsp://camera", include_bytes!("testdata/macrovideo_describe.txt")).unwrap();
+        assert_eq!(p.streams.len(), 1);
+    }
 }